Put shared popover style into a helper function
